Implement fast BFS for WordNet, WordNetFastBFS. The class returns all nodes...
...reachable from the starting node v. This eliminates the need to iterate
over the entire graph.

The second optimization is using a HashMap to store visited nodes rather
than using an array of size V and setting the visited nodes to true. For
WordNet, the number of visited nodes is very small (5-20) compared to
V (>80,000). So initializing the array will take time while only very few
elements will ever be updated as visited. HashMap will be very efficient
in this case.